backend/pyproject.toml
[tool.poetry]
authors = ["AlanTheKnight <51376126+AlanTheKnight@users.noreply.github.com>"]
description = "Open-source digital diary for schools made with Django & Vue.JS"
license = "GNU GPLv3"
name = "diary56x"
version = "0.0.1"

[tool.poetry.dependencies]
Django = "^3.2.9"
Markdown = "^3.3.4"
Pillow = "^9.0.1"
django-cleanup = "^5.1.0"
django-cors-headers = "^3.10.0"
django-debug-toolbar = "^3.2"
django-extensions = "^3.1.2"
django-filter = "^2.4.0"
django-widget-tweaks = "^1.4.8"
djangorestframework = "^3.12.4"
mkdocs = "^1.2"
poetry = "^1.1.6"
python = "^3.9"
sorl-thumbnail-serializer-field = "^0.2.1"
toml = "^0.10.2"
whitenoise = "^5.3.0"
djoser = "^2.1.0"
mkdocs-material = "^8.2.9"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
black = "22.3.0"
flake8 = "^4.0.1"
flake8-builtins = "^1.5.3"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
flake8-quotes = "^3.3.1"
isort = "^5.10.1"
mkdocs = "^1.2.3"
pylint = "^2.11.1"
pylint-django = "^2.4.4"
pytest = "^6.2.5"
pytest-django = "^4.4.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
line-length = 99
